Ranking messages added

+++ src/main/java/com/appspot/sportsrivals/model/Cities.java
@@ -6,12 +6,12 @@ public class Cities {
 
 	@Id private String id;
 	private String name;
-	private Integer rating;
+	private Double rating;
 	private String imageUrl;
 
 	public Cities(){};
 
-	public Cities(String name, Integer rating, String imageUrl){
+	public Cities(String name, Double rating, String imageUrl){
 		this.name = name;
 		this.rating = rating;
 		this.imageUrl = imageUrl;
@@ -19,11 +19,11 @@ public class Cities {
 
 	public String getId() { return id; }
 	public String getName(){ return name; }
-	public Integer getRating(){ return rating; }
+	public Double getRating(){ return rating; }
 	public String getImageUrl(){ return imageUrl; }
 
 	public void setName(String name){ this.name = name; }
-	public void setRating(int rating){ this.rating = rating; }
+	public void setRating(Double rating){ this.rating = rating; }
 	public void setImageUrl(String imageUrl){ this.imageUrl = imageUrl; }
 
 }
