+++ src/main/java/com/appspot/sportsrivals/model/Teams.java
@@ -9,7 +9,7 @@ public class Teams {
 
 	@Id private String id;
 	private String name;
-	private Integer rating;
+	private Double rating;
 	private String imageUrl;
 	private String sportRadarId;
 	@DBRef private Sports sport;
@@ -18,7 +18,7 @@ public class Teams {
 
 	public Teams(){}
 
-	public Teams(String name, Integer rating, String imageUrl, String sportRadarId, final Sports sport, final Cities city, final States state) {
+	public Teams(String name, Double rating, String imageUrl, String sportRadarId, final Sports sport, final Cities city, final States state) {
 		this.setName(name);
 		this.setRating(rating);
 		this.setImageUrl(imageUrl);
@@ -30,7 +30,7 @@ public class Teams {
 
 	public String getId() { return id; }
 	public String getName() { return name; }
-	public Integer getRating() { return rating; }
+	public Double getRating() { return rating; }
 	public String getImageUrl() { return imageUrl; }
 	public String getSportRadarId() { return sportRadarId; }
 	public Sports getSport() { return sport; }
@@ -39,7 +39,7 @@ public class Teams {
 
 	public void setId(String id) { this.id = id; };
 	public void setName(String name) { this.name = name; }
-	public void setRating(Integer rating) { this.rating = rating; }
+	public void setRating(Double rating) { this.rating = rating; }
 	public void setImageUrl(String imageUrl) { this.imageUrl = imageUrl; }
 	public void setSportRadarId(String sportRadarId) { this.sportRadarId = sportRadarId; }
 	public void setSport(Sports sport) { this.sport = sport; }
